Added debug dialog, cleaned not needed files, change architecture for debug stuff
