+++ app/src/main/java/com/tari/android/wallet/ui/common/CommonViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.ViewModel
 import com.orhanobut.logger.Logger
 import com.orhanobut.logger.Printer
 import com.tari.android.wallet.application.WalletState
+import com.tari.android.wallet.data.sharedPrefs.network.NetworkRepository
 import com.tari.android.wallet.di.ApplicationComponent
 import com.tari.android.wallet.di.DiContainer
 import com.tari.android.wallet.event.EventBus
@@ -26,6 +27,9 @@ open class CommonViewModel : ViewModel() {
 
     @Inject
     lateinit var resourceManager: ResourceManager
+
+    @Inject
+    lateinit var networkRepository: NetworkRepository
 
     val logger: Printer
         get() = Logger.t("screen").t(this::class.simpleName)
