+++ app/src/main/java/com/tari/android/wallet/ui/dialog/modular/ModularDialog.kt
@@ -20,6 +20,8 @@ import com.tari.android.wallet.ui.dialog.modular.modules.customBaseNodeBody.Cust
 import com.tari.android.wallet.ui.dialog.modular.modules.head.*
 import com.tari.android.wallet.ui.dialog.modular.modules.imageModule.ImageModule
 import com.tari.android.wallet.ui.dialog.modular.modules.imageModule.ImageModuleView
+import com.tari.android.wallet.ui.dialog.modular.modules.option.OptionModule
+import com.tari.android.wallet.ui.dialog.modular.modules.option.OptionModuleView
 import com.tari.android.wallet.ui.fragment.send.addAmount.feeModule.FeeModule
 import com.tari.android.wallet.ui.fragment.send.addAmount.feeModule.FeeModuleView
 import com.tari.android.wallet.ui.fragment.send.shareQr.ShareQRCodeModuleView
@@ -65,6 +67,7 @@ open class ModularDialog(val context: Context) : TariDialog {
                 is HeadBoldSpannableModule -> HeadBoldSpannableModuleView(context, module)
                 is ImageModule -> ImageModuleView(context, module)
                 is BodyModule -> BodyModuleView(context, module)
+                is OptionModule -> OptionModuleView(context, module)
                 is ButtonModule -> ButtonModuleView(context, module) { dialog.dismiss() }
                 is CustomBaseNodeBodyModule -> CustomBaseNodeBodyModuleView(context, module)
                 is ShareQrCodeModule -> ShareQRCodeModuleView(context, module)
