+++ app/src/main/java/com/tari/android/wallet/ui/dialog/modular/modules/option/OptionModule.kt
@@ -1,0 +1,5 @@
+package com.tari.android.wallet.ui.dialog.modular.modules.option
+
+import com.tari.android.wallet.ui.dialog.modular.IDialogModule
+
+class OptionModule(val text: String? = null, action: () -> Unit = {}) : IDialogModule()
