+++ app/src/main/java/com/tari/android/wallet/ui/fragment/auth/AuthViewModel.kt
@@ -13,9 +13,6 @@ class AuthViewModel : CommonViewModel() {
     lateinit var sharedPrefsWrapper: SharedPrefsRepository
 
     @Inject
-    lateinit var networkRepository: NetworkRepository
-
-    @Inject
     lateinit var authService: BiometricAuthenticationService
 
     @Inject
