--- app/src/main/java/com/tari/android/wallet/ui/fragment/debug/baseNodeConfig/BaseNodeConfigNavigation.kt
@@ -1,7 +1,0 @@
-package com.tari.android.wallet.ui.fragment.debug.baseNodeConfig
-
-sealed class BaseNodeConfigNavigation {
-    object ToAddCustomBaseNode : BaseNodeConfigNavigation()
-
-    object ToChangeBaseNode : BaseNodeConfigNavigation()
-}
