--- app/src/main/java/com/tari/android/wallet/ui/fragment/debug/baseNodeConfig/BaseNodeConfigRouter.kt
@@ -1,7 +1,0 @@
-package com.tari.android.wallet.ui.fragment.debug.baseNodeConfig
-
-interface BaseNodeConfigRouter {
-    fun toChangeBaseNode()
-
-    fun toAddCustomBaseNode()
-}
