+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/baseNodeConfig/BaseNodeRouter.kt
@@ -1,0 +1,5 @@
+package com.tari.android.wallet.ui.fragment.settings.baseNodeConfig
+
+interface BaseNodeRouter {
+    fun toAddCustomBaseNode()
+}
