+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/baseNodeConfig/changeBaseNode/adapter/ChangeBaseNodeAdapter.kt
@@ -1,0 +1,9 @@
+package com.tari.android.wallet.ui.fragment.settings.baseNodeConfig.changeBaseNode.adapter
+
+import com.tari.android.wallet.ui.common.recyclerView.CommonAdapter
+import com.tari.android.wallet.ui.common.recyclerView.CommonViewHolderItem
+import com.tari.android.wallet.ui.common.recyclerView.ViewHolderBuilder
+
+class ChangeBaseNodeAdapter : CommonAdapter<CommonViewHolderItem>() {
+    override var viewHolderBuilders: List<ViewHolderBuilder> = listOf(BaseNodeViewHolder.getBuilder())
+}
