+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/baseNodeConfig/validator/Validator.kt
@@ -1,0 +1,11 @@
+package com.tari.android.wallet.ui.fragment.settings.baseNodeConfig.validator
+
+interface Validator {
+    fun validate(string: String) : State
+
+    sealed class State {
+        object Valid: State()
+        object Invalid : State()
+        object Neutral: State()
+    }
+}
