+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/logs/debugLog/adapter/LogFileSpinnerAdapter.kt
@@ -1,0 +1,73 @@
+/**
+ * Copyright 2020 The Tari Project
+ *
+ * Redistribution and use in source and binary forms, with or
+ * without modification, are permitted provided that the
+ * following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above
+ * copyright notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of
+ * its contributors may be used to endorse or promote products
+ * derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
+ * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+ * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package com.tari.android.wallet.ui.fragment.settings.logs.debugLog.adapter
+
+import android.content.Context
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.BaseAdapter
+import android.widget.SpinnerAdapter
+import android.widget.TextView
+import com.tari.android.wallet.R
+import com.tari.android.wallet.ui.extension.gone
+import java.io.File
+
+class LogFileSpinnerAdapter(context: Context, files: List<File>) : BaseAdapter(), SpinnerAdapter {
+
+    private val inflater: LayoutInflater = LayoutInflater.from(context)
+    private val fileNames = files.map { it.name }
+
+    override fun getCount(): Int = fileNames.size
+
+    override fun getItem(position: Int): Any = fileNames[position]
+
+    override fun getItemId(position: Int): Long = position.toLong()
+
+    override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View {
+        val view = convertView ?: inflater.inflate(R.layout.item_log_file_spinner, null)
+        view.findViewById<TextView>(R.id.log_file_spinner_item_txt_file_name).text = fileNames[position]
+        view.findViewById<View>(R.id.log_file_spinner_item_vw_gray_bg).gone()
+        return view
+    }
+
+    override fun getDropDownView(position: Int, convertView: View?, parent: ViewGroup?): View {
+        val view = getView(position, convertView, parent)
+        val bgView = view.findViewById<View>(R.id.log_file_spinner_item_vw_gray_bg)
+        bgView.visibility = when (position % 2) {
+            0 -> View.GONE
+            else -> View.VISIBLE
+        }
+        return view
+    }
+}
