+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/networkSelection/NetworkSelectionViewModel.kt
@@ -21,9 +21,6 @@ import javax.inject.Inject
 class NetworkSelectionViewModel : CommonViewModel() {
 
     @Inject
-    lateinit var networkRepository: NetworkRepository
-
-    @Inject
     lateinit var walletConfig: WalletConfig
 
     @Inject
