--- app/src/main/res/layout/fragment_about.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:orientation="vertical">
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@null"
-        android:layout_marginTop="-50dp"
-        android:alpha="0.9"
-        android:src="@drawable/about_logo" />
-
-    <com.tari.android.wallet.ui.component.CustomFontTextView
-        android:id="@+id/version_name_text_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="18dp"
-        android:textColor="@color/black"
-        android:alpha="0.8"
-        android:textSize="16sp"
-        app:customFont="@string/font_regular"
-        tools:text="v0.4.1" />
-
-    <com.tari.android.wallet.ui.component.CustomFontTextView
-        android:id="@+id/version_code_text_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:textColor="@color/black"
-        android:alpha="0.8"
-        android:textSize="15sp"
-        app:customFont="@string/font_regular"
-        tools:text="b123" />
-
-</LinearLayout>
